refactor(kernel): extract LCG + pickTileValue into _internal.ts [1.1]
Phase 1.1 of the perf plan. Pure refactor — the same three private
functions (lcgNext, lcgFloat, pickTileValue) were defined twice,
once in board.ts and once in index.ts. Moving them to a single
internal module gives us one place to apply the actual perf
changes that follow (1.2 caches the spawn-weight CDF here).

No behavior change. Tests, typecheck, boundaries: all green.

Bench delta vs Phase 0 baseline:
  applyAction         108,158 → 107,878 hz  (within noise)
  full game seed 1    112.86  → 118.47  hz  (within noise)
  full game seed 2    68.36   → 69.57   hz  (within noise)

- src/game-kernel/_internal.ts: new home for LCG + pickTileValue
- src/game-kernel/board.ts: import from _internal, drop duplicates
- src/game-kernel/index.ts: import from _internal, drop duplicates
